fix: fix pr tests
