+++ packages/mwp-app-render/src/components/BrowserApp.test.jsx
@@ -2,6 +2,10 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import BrowserApp from './BrowserApp';
 
+jest.mock('./ApolloProvider', () => {
+	return props => <div {...props} />;
+});
+
 // TODO: skipping test until browser test environment enabled
 describe.skip('BrowserApp', function() {
 	const routes = [];
