+++ packages/mwp-app-render/src/components/ServerApp.test.jsx
@@ -2,6 +2,10 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import ServerApp from './ServerApp';
 
+jest.mock('./ApolloProvider', () => {
+	return props => <div {...props} />;
+});
+
 describe('ServerApp', function() {
 	const routes = [];
 	const basename = '';
